requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0
google-generativeai>=0.8,<0.9
python-dotenv>=1.0,<1.1
